Separate app bootstrap wiring by responsibility
Split the app package bootstrap code into files for config and database setup, storage wiring, OIDC provider construction, route registration, and server lifecycle. Keep app.Run as the only public entrypoint so behavior stays the same while the wiring boundaries are reviewable.

Constraint: This keeps zitadel/oidc storage in one package because the concrete Storage type must satisfy op.Storage and DeviceAuthorizationStorage.

Rejected: Splitting internal/storage into auth/device subpackages | the OIDC library expects one concrete storage implementation.

Confidence: high

Scope-risk: narrow

Directive: Add new bootstrap concerns to the matching internal/app file instead of growing app.go again.

Tested: git diff --cached --check; go test ./cmd/authgate ./internal/app; go test ./...; go vet ./...

Not-tested: Live server startup against a production database.
